+++ src/private/update/InvokeNovaUpdateNotificationPreferenceChange.ps1
@@ -1,0 +1,9 @@
+function Invoke-NovaUpdateNotificationPreferenceChange {
+    [CmdletBinding()]
+    param(
+        [Parameter(Mandatory)][pscustomobject]$WorkflowContext
+    )
+
+    Write-NovaUpdateNotificationPreference -PrereleaseNotificationsEnabled:$WorkflowContext.PrereleaseNotificationsEnabled
+    return Get-NovaUpdateNotificationPreferenceStatus
+}